Massive overhaul that enables conversion of PDF files to markdown format, including tables, using the JSON provided by the Adobe API.
Instead of creating the tables from the data within the JSON the ".csv" output provided by the API is used with "pandas" to convert it to markdown format.
The Path/Text pairs are then targeted and added to a list with markdown formatting applied. This still targets the table data within the JSON which is then stripped from list, leaving only one ocurrence for each table in the PDF.
The converted tables are then added to a list at the positions marked in the previous list replacing that item before everything is finally wrote out to a file.

This needs a refactor as it is very messy, but it works for now.